Сделает рефакторинг вспомогательных функций
